Enable HelloWorld in x86-64 NaCl.
This does *not* enable threading support, only make it buildable. The TLS
support was disabled by converting all thread local variables to globals
(in bss).

The change was supposed to only make runnable-ld.so print the help message, but
the requirement to have sources buildable pulled so many changes that actually
making HeloWorld run was a minor change.

   * elf/Makefile
     elf/Versions
     sysdeps/nacl/nacl_read_tp.c
       nacl_read_tp: TLS support for General Dynamic model

   * elf/dl-fini.c
       uninvestigated crasher commented out

   * elf/dl-reloc.c
     elf/rtld.c
     include/errno.h
     nptl/sysdeps/x86_64/tls.h
     sysdeps/unix/x86_64/sysdep.S
     sysdeps/x86_64/sysdep.h
       TLS_HACK: make thread-local variables global, disable TLS initialization

   * elf/dl-sysdep.c
       argc was pushed onto the stack via a 64bit push, take argv from a shifted
       place.

   * elf/nacl_fixup_ldso.py
       process ELF64 binaries as well

   * make_sysd_rules.py
       add more .c overrides

   * nacl/dyn-link/ldscripts/elf64_nacl.x
     nacl/dyn-link/ldscripts/elf64_nacl.xs
       linker scripts done in the most part by copying 32bit linker scripts

   * nptl/pthread_mutex_timedlock.c
       a temporary change that will have to be eliminated as deadcode when we
       integrate threading better

   * nptl/sysdeps/x86_64/pthreaddef.h
       tiny inline assembly fix

   * stdlib/llabs.c
     sysdeps/wordsize-64/labs.c
       imaxabs is equal to llabs since labs is smaller

   * sysdeps/generic/gmp-mparam.h
       make gmp work using chunks identical to the i386 case

   * sysdeps/generic/stdint.h
       long != long long, hence a huge change

   * sysdeps/nacl/mmap64.c
     sysdeps/nacl/open64.c
       glibc wants them supported for some reason, maybe we should kill them later

   * sysdeps/nacl/umount.c
       make-syscalls.sh failed to generate an unimplemented syscall stub for
       'umount', creating it by hand

   * sysdeps/nacl/write.c
     sysdeps/unix/sysv/linux/wordsize-64/fxstat.c

   * .../unix/sysv/linux/wordsize-64/posix_fadvise.c
       posix_fadvise commented out, should rather reuse the generic:
       sysdeps/unix/sysv/linux/posix_fadvise.c

   * sysdeps/unix/sysv/linux/x86_64/setcontext.S
       just make it compilable

   * sysdeps/unix/sysv/linux/x86_64/sysdep.h
       disable pointer mangling
       also TLS_HACK

   * sysdeps/x86_64/Makefile
     sysdeps/x86_64/fpu/bits/mathinline.h
     sysdeps/x86_64/fpu/s_nexttoward.c
       denote that we do not support long double in NaCl

   * sysdeps/x86_64/dl-machine.h
     sysdeps/x86_64/elf/initfini.c
       trivial fixes in _dl_start_user
       added support for relocation type R_X86_64_32 in dynamic linker

   * sysdeps/x86_64/dl-trampoline.S
       sandboxing fixes in assembly

BUG=http://code.google.com/p/nativeclient/issues/detail?id=940
TEST=make build-with-glibc CROSSARCH=nacl64 MULTILIB=no && make nacl-check CROSSARCH=nacl64
(sel_ldr -s would refuse to run the test though, will report the issue as soon
 as this change gets committed)

Review URL: http://codereview.chromium.org/5059002

+++ sysdeps/nacl/write.c
@@ -6,7 +6,7 @@
 #include <nacl_syscalls.h>
 
 
-int __write(int desc, void const *buf, size_t count)
+ssize_t __write(int desc, void const *buf, size_t count)
 {
   int result = NACL_SYSCALL (write) (desc, buf, count);
   if (result < 0) {
